+++ apps/block_scout_web/assets/js/pages/verified_contracts.js
@@ -59,7 +59,7 @@ if ($('[data-page="verified-contracts-list"]').length) {
   const $element = $('[data-async-listing]')
 
   $element.on('click', '[data-next-page-button], [data-prev-page-button]', (event) => {
-    document.getElementById('verified-contratc-list').scrollIntoView()
+    document.getElementById('verified-contracts-list').scrollIntoView()
   })
 
   const store = createAsyncLoadStore(reducer, initialState, 'dataset.identifierHash')