+++ apps/explorer/test/explorer/counters/verified_contracts_counter_test.exs
@@ -1,9 +1,8 @@
 defmodule Explorer.Counters.VerifiedContractsCounterTest do
   use Explorer.DataCase
 
-  alias Explorer.Counters.VerifiedContractsCounter
+  alias Explorer.Chain.Cache.VerifiedContractsCounter
   alias Explorer.Chain
-  alias Explorer.Chain.Transaction
 
   test "populates the cache with the number of verified contracts" do
     insert(:smart_contract)
@@ -13,6 +12,6 @@ defmodule Explorer.Counters.VerifiedContractsCounterTest do
     start_supervised!(VerifiedContractsCounter)
     VerifiedContractsCounter.consolidate()
 
-    assert Chain.count_verified_contracts_from_cache() == 3
+    assert Chain.count_verified_contracts_from_cache() == Decimal.new(3)
   end
 end