+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 
+- [#6324](https://github.com/blockscout/blockscout/pull/6324) - Add verified contracts list page
 - [#6351](https://github.com/blockscout/blockscout/pull/6351) - Enable forum link env var
 - [#6316](https://github.com/blockscout/blockscout/pull/6316) - Copy public tags functionality to master
 - [#6196](https://github.com/blockscout/blockscout/pull/6196) - INDEXER_CATCHUP_BLOCKS_BATCH_SIZE and INDEXER_CATCHUP_BLOCKS_CONCURRENCY env varaibles